Port the SDL examples to SDL2
If we only want a simple solution to port examples from SDL1.2 to SDL2, we can
keep using the remaining API between SDL1.2 and SDL2 and exchanging the expired API directly.

However, I am not sure if this is a good solution because it keeps using
the 'SDL_Surface' object. I feel like according to the
[SDL 1.2 to 2.0 Migration Guide](https://wiki.libsdl.org/MigrationGuide), we
don't need to work on the 'SDL_Surface' object but could still do the same result of
examples. It means that we could use the SDL2 APIs more like the document shows us.

But maybe we can just enable to use SDL2 now and refine those examples in the future.